Make zip package checksum predictable
When packaging files, Zip adds some extra information along with the
modification time of each file and folder to the output archive. The
effect is that the checksum of the output archive changes every time you
compile your handler. Even if the checksum of the output handler.so
binary hasn't changed. This is very annoying particulary when using aws
cloudformation package command in a multi step ci/cd pipeline.
The proposed solution sets a fixed modification time for all generated
files and removes zip extra information. Thus, the checksum of the
archive will be modified if and only if the checksums of contained files
change.

CAT: #feat