Make build and clean processes less annoying
Previously, the current working directory was mounted in the /tmp folder
of the Docker process. As go uses the /tmp folder to create intermediate
build artifacts, one could see some strange folders
appearing/disappearing during the build process. This patch change the
mounted folder to /build to prevent such behaviors. This is not a
breaking change.
Moreover, the $(HANDLER) folder is removed from the `clean` target of
the Makefile to prevent unexpected removal of such folders containing
user files. In fact, this folder is never generated by the build process
inside the user working directory and should not be targeted for
deletion by the Makefile.

CAT: #feat