Fix bug with multiple GOPATH
GOPATH may be a list of directories and this patch adds the possibility
to support such cases.
It should not break established build processes relying on the example
Makefile but be aware that mounted volumes are now slightly diffrent, so
if you've made strong assumption on them, you have to update your build
process. It's also worth noting that the previous generation of the
example Makefile will continue to work as there is no modification on
the Docker image. This new generation adds the ability to support
multiple GOPATH.

CAT: #fix
REF: #25