Add :Z to Docker volume mounts for SELinux
When running the sample Makefile on a host with SELinux enabled (Fedora
27 in my case), the Docker container is not able to access the files
that are mapped into the container via the volumes. This is due to
SELinux on the host not having the files labeled correctly. Adding
a :Z to the end of the volume mounts will correct this issue. See
https://www.projectatomic.io/blog/2016/03/dwalsh_selinux_containers/
for more details about the :Z and :z options that are available. These
options are silently ignored for hosts that do not have SELinux enabled
(including OS/X).

CAT: #feat
REF: #46
THX: @masneyb